test: extract domain-specific tests into dedicated test files
Move tests from the monolithic organization_test.go into
dedicated test files matching their source files:
application_test.go, auto_prune_test.go, org_marketplace_test.go,
org_robot_test.go, org_team_test.go, proxy_cache_test.go, and
quota_test.go. Rename team_test.go to org_team_test.go.

Every lib source file now has a corresponding test file.

Closes #171
